requirements.txt
pandas
requests
tqdm
plotly
python-dotenv
